cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fhe-gbdt-cli"
version = "1.0.0"
description = "CLI tool for FHE-GBDT-Serving platform - encrypted machine learning inference"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "FHE-GBDT Team", email = "team@fhe-gbdt.io"}
]
keywords = [
    "fhe",
    "fully-homomorphic-encryption",
    "gbdt",
    "machine-learning",
    "privacy",
    "cli"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Security :: Cryptography",
]
dependencies = [
    "click>=8.0.0",
    "rich>=13.0.0",
    "httpx>=0.24.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
fhe-gbdt = "fhe_gbdt_cli.main:cli"

[project.urls]
Homepage = "https://github.com/fhe-gbdt/fhe-gbdt-serving"
Documentation = "https://docs.fhe-gbdt.io"
Repository = "https://github.com/fhe-gbdt/fhe-gbdt-serving"
Issues = "https://github.com/fhe-gbdt/fhe-gbdt-serving/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["fhe_gbdt_cli*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
